feat: support textarea smarthint
