+++ packages/toolkit/src/lib/use-instill-form/components/smart-hint/index.ts
@@ -1,0 +1,7 @@
+import { TextArea } from "./TextArea";
+import { TextField } from "./TextField";
+
+export const SmartHintFields = {
+  TextArea,
+  TextField,
+};
