+++ packages/toolkit/src/lib/use-instill-form/type.ts
@@ -13,6 +13,7 @@ export type InstillCustomProps = {
   instillAcceptFormats?: string[];
   instillShortDescription?: string;
   instillUIOrder?: number;
+  instillUiMultiline?: boolean;
   instillEditOnNodeFields?: string[];
   instillCredentialField?: boolean;
 };
